IGNITE-21240 Removed deprecated authorization methods from Security Context.
