+++ modules/core/src/test/java/org/apache/ignite/internal/processors/security/impl/TestSecurityContext.java
@@ -18,10 +18,7 @@
 package org.apache.ignite.internal.processors.security.impl;
 
 import java.io.Serializable;
-import java.util.Collection;
 import org.apache.ignite.internal.processors.security.SecurityContext;
-import org.apache.ignite.internal.util.typedef.F;
-import org.apache.ignite.plugin.security.SecurityPermission;
 import org.apache.ignite.plugin.security.SecuritySubject;
 
 /**
@@ -38,90 +35,11 @@ public class TestSecurityContext implements SecurityContext, Serializable {
         this.subject = subject;
     }
 
-    /**
-     * @param opName Op name.
-     * @param perm Permission.
-     */
-    public boolean operationAllowed(String opName, SecurityPermission perm) {
-        switch (perm) {
-            case CACHE_CREATE:
-            case CACHE_DESTROY:
-                return systemOperationAllowed(perm) || cacheOperationAllowed(opName, perm);
-
-            case CACHE_PUT:
-            case CACHE_READ:
-            case CACHE_REMOVE:
-                return cacheOperationAllowed(opName, perm);
-
-            case TASK_CANCEL:
-            case TASK_EXECUTE:
-                return taskOperationAllowed(opName, perm);
-
-            case SERVICE_DEPLOY:
-            case SERVICE_INVOKE:
-            case SERVICE_CANCEL:
-                return serviceOperationAllowed(opName, perm);
-
-            case EVENTS_DISABLE:
-            case EVENTS_ENABLE:
-            case ADMIN_VIEW:
-            case ADMIN_CACHE:
-            case ADMIN_QUERY:
-            case ADMIN_OPS:
-            case ADMIN_SNAPSHOT:
-            case ADMIN_CLUSTER_STATE:
-            case JOIN_AS_SERVER:
-            case ADMIN_KILL:
-            case ADMIN_USER_ACCESS:
-            case ADMIN_CLUSTER_NODE_STOP:
-            case ADMIN_CLUSTER_NODE_START:
-                return systemOperationAllowed(perm);
-
-            default:
-                throw new IllegalStateException("Invalid security permission: " + perm);
-        }
-    }
-
     /** {@inheritDoc} */
     @Override public SecuritySubject subject() {
         return subject;
     }
 
-    /** {@inheritDoc} */
-    @Override public boolean taskOperationAllowed(String taskClsName, SecurityPermission perm) {
-        return hasPermission(subject.permissions().taskPermissions().get(taskClsName), perm);
-    }
-
-    /** {@inheritDoc} */
-    @Override public boolean cacheOperationAllowed(String cacheName, SecurityPermission perm) {
-        return hasPermission(subject.permissions().cachePermissions().get(cacheName), perm);
-    }
-
-    /** {@inheritDoc} */
-    @Override public boolean serviceOperationAllowed(String srvcName, SecurityPermission perm) {
-        return hasPermission(subject.permissions().servicePermissions().get(srvcName), perm);
-    }
-
-    /** {@inheritDoc} */
-    @Override public boolean systemOperationAllowed(SecurityPermission perm) {
-        Collection<SecurityPermission> perms = subject.permissions().systemPermissions();
-
-        if (F.isEmpty(perms))
-            return subject.permissions().defaultAllowAll();
-
-        return perms.stream().anyMatch(p -> perm == p);
-    }
-
-    /**
-     * @param perms Permissions.
-     * @param perm Permission.
-     */
-    private boolean hasPermission(Collection<SecurityPermission> perms, SecurityPermission perm) {
-        if (perms == null)
-            return subject.permissions().defaultAllowAll();
-
-        return perms.stream().anyMatch(p -> perm == p);
-    }
 
     /** {@inheritDoc} */
     @Override public String toString() {
