+++ modules/core/src/test/java/org/apache/ignite/internal/processors/security/impl/TestSecuritySubject.java
@@ -21,7 +21,6 @@ import java.net.InetSocketAddress;
 import java.security.PermissionCollection;
 import java.security.cert.Certificate;
 import java.util.UUID;
-import org.apache.ignite.plugin.security.SecurityPermissionSet;
 import org.apache.ignite.plugin.security.SecuritySubject;
 import org.apache.ignite.plugin.security.SecuritySubjectType;
 
@@ -41,37 +40,11 @@ public class TestSecuritySubject implements SecuritySubject {
     /** Address. */
     private InetSocketAddress addr;
 
-    /** Permissions. */
-    private SecurityPermissionSet perms;
-
     /** Permissions for Sandbox checks. */
     private PermissionCollection sandboxPerms;
 
     /** Client certificates. */
     private Certificate[] certs;
-
-    /**
-     * Default constructor.
-     */
-    public TestSecuritySubject() {
-        // No-op.
-    }
-
-    /**
-     * @param id Id.
-     * @param login Login.
-     * @param addr Address.
-     * @param perms Permissions.
-     */
-    public TestSecuritySubject(UUID id,
-        Object login,
-        InetSocketAddress addr,
-        SecurityPermissionSet perms) {
-        this.id = id;
-        this.login = login;
-        this.addr = addr;
-        this.perms = perms;
-    }
 
     /** {@inheritDoc} */
     @Override public UUID id() {
@@ -130,20 +103,6 @@ public class TestSecuritySubject implements SecuritySubject {
     }
 
     /** {@inheritDoc} */
-    @Override public SecurityPermissionSet permissions() {
-        return perms;
-    }
-
-    /**
-     * @param perms Permissions.
-     */
-    public TestSecuritySubject setPerms(SecurityPermissionSet perms) {
-        this.perms = perms;
-
-        return this;
-    }
-
-    /** {@inheritDoc} */
     @Override public PermissionCollection sandboxPermissions() {
         return sandboxPerms;
     }
@@ -160,15 +119,12 @@ public class TestSecuritySubject implements SecuritySubject {
         return certs;
     }
 
-    /**
-     * @param perms Permissions.
-     */
+    /** */
     public TestSecuritySubject setCerts(Certificate[] certs) {
         this.certs = certs;
 
         return this;
     }
-
 
     /** {@inheritDoc} */
     @Override public String toString() {
