+++ modules/core/src/test/java/org/apache/ignite/spi/discovery/tcp/TestReconnectProcessor.java
@@ -33,7 +33,6 @@ import org.apache.ignite.plugin.security.AuthenticationContext;
 import org.apache.ignite.plugin.security.SecurityCredentials;
 import org.apache.ignite.plugin.security.SecurityException;
 import org.apache.ignite.plugin.security.SecurityPermission;
-import org.apache.ignite.plugin.security.SecurityPermissionSet;
 import org.apache.ignite.plugin.security.SecuritySubject;
 import org.apache.ignite.plugin.security.SecuritySubjectType;
 import org.jetbrains.annotations.Nullable;
@@ -140,11 +139,6 @@ public class TestReconnectProcessor extends GridProcessorAdapter implements Grid
         @Override public InetSocketAddress address() {
             return null;
         }
-
-        /** {@inheritDoc} */
-        @Override public SecurityPermissionSet permissions() {
-            return null;
-        }
     }
 
     /**
@@ -167,26 +161,6 @@ public class TestReconnectProcessor extends GridProcessorAdapter implements Grid
         /** {@inheritDoc} */
         @Override public SecuritySubject subject() {
             return subj;
-        }
-
-        /** {@inheritDoc} */
-        @Override public boolean taskOperationAllowed(String taskClsName, SecurityPermission perm) {
-            return true;
-        }
-
-        /** {@inheritDoc} */
-        @Override public boolean cacheOperationAllowed(String cacheName, SecurityPermission perm) {
-            return true;
-        }
-
-        /** {@inheritDoc} */
-        @Override public boolean serviceOperationAllowed(String srvcName, SecurityPermission perm) {
-            return true;
-        }
-
-        /** {@inheritDoc} */
-        @Override public boolean systemOperationAllowed(SecurityPermission perm) {
-            return true;
         }
     }
 }
